+++ src/main/java/fr/alasdiablo/mods/resources/ExtendedResourcesRegistries.java
@@ -1,37 +1,56 @@
 package fr.alasdiablo.mods.resources;
 
 public class ExtendedResourcesRegistries {
-    public static final String COAL_DUST    = "coal_dust";
-    public static final String COPPER_DUST  = "copper_dust";
-    public static final String DIAMOND_DUST = "diamond_dust";
-    public static final String EMERALD_DUST = "emerald_dust";
-    public static final String GOLD_DUST    = "gold_dust";
-    public static final String IRON_DUST    = "iron_dust";
-    public static final String LAPIS_DUST   = "lapis_dust";
+    // Ingots
+    public static final String STEEL_INGOT = "steel_ingot";
 
-    public static final String COAL_NUGGET     = "coal_nugget";
-    public static final String COPPER_NUGGET   = "copper_nugget";
-    public static final String DIAMOND_NUGGET  = "diamond_nugget";
-    public static final String EMERALD_NUGGET  = "emerald_nugget";
-    public static final String LAPIS_NUGGET    = "lapis_nugget";
-    public static final String REDSTONE_NUGGET = "redstone_nugget";
+    // Dusts
+    public static final String COAL_DUST          = "coal_dust";
+    public static final String COPPER_DUST        = "copper_dust";
+    public static final String DIAMOND_DUST       = "diamond_dust";
+    public static final String EMERALD_DUST       = "emerald_dust";
+    public static final String GOLD_DUST          = "gold_dust";
+    public static final String IRON_DUST          = "iron_dust";
+    public static final String LAPIS_DUST         = "lapis_dust";
+    public static final String ENRICHED_IRON_DUST = "enriched_iron_dust";
+    public static final String STEEL_DUST         = "steel_dust";
 
+    // Nuggets
+    public static final String COAL_NUGGET        = "coal_nugget";
+    public static final String COPPER_NUGGET      = "copper_nugget";
+    public static final String DIAMOND_NUGGET     = "diamond_nugget";
+    public static final String EMERALD_NUGGET     = "emerald_nugget";
+    public static final String LAPIS_NUGGET       = "lapis_nugget";
+    public static final String REDSTONE_NUGGET    = "redstone_nugget";
     public static final String NETHERITE_NUGGET   = "netherite_nugget";
     public static final String NETHERITE_FRAGMENT = "netherite_fragment";
+    public static final String STEEL_NUGGET       = "steel_nugget";
 
+    // Plates
     public static final String COPPER_PLATE = "copper_plate";
     public static final String IRON_PLATE   = "iron_plate";
     public static final String GOLD_PLATE   = "gold_plate";
+    public static final String STEEL_PLATE  = "steel_plate";
 
+    // Gears
     public static final String WOODEN_GEAR  = "wooden_gear";
     public static final String COPPER_GEAR  = "copper_gear";
     public static final String IRON_GEAR    = "iron_gear";
     public static final String GOLD_GEAR    = "gold_gear";
+    public static final String STEEL_GEAR   = "steel_gear";
 
+    // Rods
     public static final String COPPER_ROD  = "copper_rod";
     public static final String IRON_ROD    = "iron_rod";
     public static final String GOLD_ROD    = "gold_rod";
+    public static final String STEEL_ROD   = "steel_rod";
 
-    public static final String METAL_HAMMER = "metal_hammer";
-    public static final String METAL_CUTTER = "metal_cutter";
+    // Tools part
+    public static final String MORTAR = "mortar";
+    public static final String PESTLE = "pestle";
+
+    // Tools
+    public static final String METAL_HAMMER      = "metal_hammer";
+    public static final String METAL_CUTTER      = "metal_cutter";
+    public static final String MORTAR_AND_PESTLE = "mortar_and_pestle";
 }