+++ src/main/java/fr/alasdiablo/mods/resources/data/lang/LanguagesProvider.java
@@ -27,6 +27,7 @@ public class LanguagesProvider {
             @Override
             protected void addTranslations() {
                 super.addTranslations();
+                this.add(ExtendedResourcesItems.STEEL_INGOT.get(), "Steel Ingot");
 
                 this.add(ExtendedResourcesItems.COAL_DUST.get(), "Coal Dust");
                 this.add(ExtendedResourcesItems.COPPER_DUST.get(), "Copper Dust");
@@ -35,6 +36,8 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.GOLD_DUST.get(), "Gold Dust");
                 this.add(ExtendedResourcesItems.IRON_DUST.get(), "Iron Dust");
                 this.add(ExtendedResourcesItems.LAPIS_DUST.get(), "Lapis Dust");
+                this.add(ExtendedResourcesItems.ENRICHED_IRON_DUST.get(), "Enriched Iron Dust");
+                this.add(ExtendedResourcesItems.STEEL_DUST.get(), "Steel Dust");
 
                 this.add(ExtendedResourcesItems.COAL_NUGGET.get(), "Coal Chunk");
                 this.add(ExtendedResourcesItems.COPPER_NUGGET.get(), "Copper Nugget");
@@ -42,6 +45,7 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.EMERALD_NUGGET.get(), "Emerald Shard");
                 this.add(ExtendedResourcesItems.LAPIS_NUGGET.get(), "Lapis Lazuli Fragment");
                 this.add(ExtendedResourcesItems.REDSTONE_NUGGET.get(), "Tiny Pile of Redstone");
+                this.add(ExtendedResourcesItems.STEEL_NUGGET.get(), "Steel Nugget");
 
                 this.add(ExtendedResourcesItems.NETHERITE_NUGGET.get(), "Netherite Nugget");
                 this.add(ExtendedResourcesItems.NETHERITE_FRAGMENT.get(), "Netherite Fragment");
@@ -49,18 +53,24 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.COPPER_PLATE.get(), "Copper Plate");
                 this.add(ExtendedResourcesItems.IRON_PLATE.get(), "Iron Plate");
                 this.add(ExtendedResourcesItems.GOLD_PLATE.get(), "Golden Plate");
+                this.add(ExtendedResourcesItems.STEEL_PLATE.get(), "Steel Plate");
 
                 this.add(ExtendedResourcesItems.WOODEN_GEAR.get(), "Wooden Gear");
                 this.add(ExtendedResourcesItems.COPPER_GEAR.get(), "Copper Gear");
                 this.add(ExtendedResourcesItems.IRON_GEAR.get(), "Iron Gear");
                 this.add(ExtendedResourcesItems.GOLD_GEAR.get(), "Golden Gear");
+                this.add(ExtendedResourcesItems.STEEL_GEAR.get(), "Steel Gear");
 
                 this.add(ExtendedResourcesItems.COPPER_ROD.get(), "Copper Rod");
                 this.add(ExtendedResourcesItems.IRON_ROD.get(), "Iron Rod");
                 this.add(ExtendedResourcesItems.GOLD_ROD.get(), "Golden Rod");
+                this.add(ExtendedResourcesItems.STEEL_ROD.get(), "Steel Rod");
 
                 this.add(ExtendedResourcesItems.METAL_HAMMER.get(), "Forging hammer");
                 this.add(ExtendedResourcesItems.METAL_CUTTER.get(), "Metal cutters");
+                this.add(ExtendedResourcesItems.MORTAR.get(), "Mortar");
+                this.add(ExtendedResourcesItems.PESTLE.get(), "Pestle");
+                this.add(ExtendedResourcesItems.MORTAR_AND_PESTLE.get(), "Mortar and Pestle");
             }
         }
 
@@ -104,6 +114,7 @@ public class LanguagesProvider {
             @Override
             protected void addTranslations() {
                 super.addTranslations();
+                this.add(ExtendedResourcesItems.STEEL_INGOT.get(), "Lingot d'acier");
 
                 this.add(ExtendedResourcesItems.COAL_DUST.get(), "Poussière de charbon");
                 this.add(ExtendedResourcesItems.COPPER_DUST.get(), "Poussière de cuivre");
@@ -112,6 +123,8 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.GOLD_DUST.get(), "Poussière d'or");
                 this.add(ExtendedResourcesItems.IRON_DUST.get(), "Poussière de fer");
                 this.add(ExtendedResourcesItems.LAPIS_DUST.get(), "Poussière de lapis-lazuli");
+                this.add(ExtendedResourcesItems.ENRICHED_IRON_DUST.get(), "Poussière de fer enrichie");
+                this.add(ExtendedResourcesItems.STEEL_DUST.get(), "Poussière d'acier");
 
                 this.add(ExtendedResourcesItems.COAL_NUGGET.get(), "Morceau de charbon");
                 this.add(ExtendedResourcesItems.COPPER_NUGGET.get(), "Pépite de cuivre");
@@ -119,6 +132,7 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.EMERALD_NUGGET.get(), "Éclat d'émeraude");
                 this.add(ExtendedResourcesItems.LAPIS_NUGGET.get(), "Fragment de lapis-lazuli");
                 this.add(ExtendedResourcesItems.REDSTONE_NUGGET.get(), "Petit tas de redstone");
+                this.add(ExtendedResourcesItems.STEEL_NUGGET.get(), "Pépite d'acier");
 
                 this.add(ExtendedResourcesItems.NETHERITE_NUGGET.get(), "Pépite de netherite");
                 this.add(ExtendedResourcesItems.NETHERITE_FRAGMENT.get(), "Éclat de netherite");
@@ -126,18 +140,24 @@ public class LanguagesProvider {
                 this.add(ExtendedResourcesItems.COPPER_PLATE.get(), "Plaque de cuivre");
                 this.add(ExtendedResourcesItems.IRON_PLATE.get(), "Plaque d'fer");
                 this.add(ExtendedResourcesItems.GOLD_PLATE.get(), "Plaque d'or");
+                this.add(ExtendedResourcesItems.STEEL_PLATE.get(), "Plaque d'acier");
 
                 this.add(ExtendedResourcesItems.WOODEN_GEAR.get(), "Roue dentée en bois");
                 this.add(ExtendedResourcesItems.COPPER_GEAR.get(), "Roue dentée en cuivre");
                 this.add(ExtendedResourcesItems.IRON_GEAR.get(), "Roue dentée en fer");
                 this.add(ExtendedResourcesItems.GOLD_GEAR.get(), "Roue dentée en or");
+                this.add(ExtendedResourcesItems.STEEL_GEAR.get(), "Roue dentée en acier");
 
                 this.add(ExtendedResourcesItems.COPPER_ROD.get(), "Tige en cuivre");
                 this.add(ExtendedResourcesItems.IRON_ROD.get(), "Tige en fer");
                 this.add(ExtendedResourcesItems.GOLD_ROD.get(), "Tige en or");
+                this.add(ExtendedResourcesItems.STEEL_ROD.get(), "Tige en acier");
 
                 this.add(ExtendedResourcesItems.METAL_HAMMER.get(), "Marteau de forge");
                 this.add(ExtendedResourcesItems.METAL_CUTTER.get(), "Cisaille à métaux");
+                this.add(ExtendedResourcesItems.MORTAR.get(), "Mortier");
+                this.add(ExtendedResourcesItems.PESTLE.get(), "Pilon");
+                this.add(ExtendedResourcesItems.MORTAR_AND_PESTLE.get(), "Mortier et pilon");
             }
         }