+++ src/main/java/fr/alasdiablo/mods/resources/data/model/ModelsProvider.java
@@ -7,18 +7,17 @@ import net.minecraft.client.data.models.ItemModelGenerators;
 import net.minecraft.client.data.models.ModelProvider;
 import net.minecraft.client.data.models.model.ModelTemplates;
 import net.minecraft.data.PackOutput;
-import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.NotNull;
 
 public class ModelsProvider extends ModelProvider {
-    private static final ResourceLocation GENERATED = ResourceLocation.withDefaultNamespace("item/generated");
-
     public ModelsProvider(PackOutput output) {
         super(output, ExtendedResources.MOD_ID);
     }
 
     @Override
     protected void registerModels(@NotNull BlockModelGenerators blockModels, @NotNull ItemModelGenerators itemModels) {
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_INGOT.get(), ModelTemplates.FLAT_ITEM);
+
         itemModels.generateFlatItem(ExtendedResourcesItems.COAL_DUST.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.COPPER_DUST.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.DIAMOND_DUST.get(), ModelTemplates.FLAT_ITEM);
@@ -26,6 +25,8 @@ public class ModelsProvider extends ModelProvider {
         itemModels.generateFlatItem(ExtendedResourcesItems.GOLD_DUST.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.IRON_DUST.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.LAPIS_DUST.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.ENRICHED_IRON_DUST.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_DUST.get(), ModelTemplates.FLAT_ITEM);
 
         itemModels.generateFlatItem(ExtendedResourcesItems.COAL_NUGGET.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.COPPER_NUGGET.get(), ModelTemplates.FLAT_ITEM);
@@ -33,6 +34,7 @@ public class ModelsProvider extends ModelProvider {
         itemModels.generateFlatItem(ExtendedResourcesItems.EMERALD_NUGGET.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.LAPIS_NUGGET.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.REDSTONE_NUGGET.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_NUGGET.get(), ModelTemplates.FLAT_ITEM);
 
         itemModels.generateFlatItem(ExtendedResourcesItems.NETHERITE_NUGGET.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.NETHERITE_FRAGMENT.get(), ModelTemplates.FLAT_ITEM);
@@ -40,17 +42,24 @@ public class ModelsProvider extends ModelProvider {
         itemModels.generateFlatItem(ExtendedResourcesItems.COPPER_PLATE.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.IRON_PLATE.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.GOLD_PLATE.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_PLATE.get(), ModelTemplates.FLAT_ITEM);
 
         itemModels.generateFlatItem(ExtendedResourcesItems.WOODEN_GEAR.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.COPPER_GEAR.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.IRON_GEAR.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.GOLD_GEAR.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_GEAR.get(), ModelTemplates.FLAT_ITEM);
 
         itemModels.generateFlatItem(ExtendedResourcesItems.COPPER_ROD.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.IRON_ROD.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.GOLD_ROD.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.STEEL_ROD.get(), ModelTemplates.FLAT_ITEM);
+
+        itemModels.generateFlatItem(ExtendedResourcesItems.MORTAR.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.PESTLE.get(), ModelTemplates.FLAT_ITEM);
 
         itemModels.generateFlatItem(ExtendedResourcesItems.METAL_HAMMER.get(), ModelTemplates.FLAT_ITEM);
         itemModels.generateFlatItem(ExtendedResourcesItems.METAL_CUTTER.get(), ModelTemplates.FLAT_ITEM);
+        itemModels.generateFlatItem(ExtendedResourcesItems.MORTAR_AND_PESTLE.get(), ModelTemplates.FLAT_ITEM);
     }
 }