+++ src/main/java/fr/alasdiablo/mods/resources/data/tag/ItemsTagsProvider.java
@@ -28,6 +28,8 @@ public class ItemsTagsProvider extends ItemTagsProvider {
         this.tag(DUSTS_GOLD).add(GOLD_DUST.get());
         this.tag(DUSTS_IRON).add(IRON_DUST.get());
         this.tag(DUSTS_LAPIS).add(LAPIS_DUST.get());
+        this.tag(DUSTS_ENRICHED_IRON).add(ENRICHED_IRON_DUST.get());
+        this.tag(DUSTS_STEEL).add(STEEL_DUST.get());
 
         this.tag(Tags.Items.DUSTS).addTag(DUSTS_COAL);
         this.tag(Tags.Items.DUSTS).addTag(DUSTS_COPPER);
@@ -36,6 +38,8 @@ public class ItemsTagsProvider extends ItemTagsProvider {
         this.tag(Tags.Items.DUSTS).addTag(DUSTS_GOLD);
         this.tag(Tags.Items.DUSTS).addTag(DUSTS_IRON);
         this.tag(Tags.Items.DUSTS).addTag(DUSTS_LAPIS);
+        this.tag(Tags.Items.DUSTS).addTag(DUSTS_ENRICHED_IRON);
+        this.tag(Tags.Items.DUSTS).addTag(DUSTS_STEEL);
 
         this.tag(NUGGETS_COAL).add(COAL_NUGGET.get());
         this.tag(NUGGETS_COPPER).add(COPPER_NUGGET.get());
@@ -44,6 +48,7 @@ public class ItemsTagsProvider extends ItemTagsProvider {
         this.tag(NUGGETS_LAPIS).add(LAPIS_NUGGET.get());
         this.tag(NUGGETS_REDSTONE).add(REDSTONE_NUGGET.get());
         this.tag(NUGGETS_NETHERITE).add(NETHERITE_NUGGET.get());
+        this.tag(NUGGETS_STEEL).add(STEEL_NUGGET.get());
 
         this.tag(TINY_DUSTS_REDSTONE).add(REDSTONE_NUGGET.get());
         this.tag(TINY_DUSTS).addTag(TINY_DUSTS_REDSTONE);
@@ -55,31 +60,38 @@ public class ItemsTagsProvider extends ItemTagsProvider {
         this.tag(Tags.Items.NUGGETS).addTag(NUGGETS_LAPIS);
         this.tag(Tags.Items.NUGGETS).addTag(NUGGETS_REDSTONE);
         this.tag(Tags.Items.NUGGETS).addTag(NUGGETS_NETHERITE);
+        this.tag(Tags.Items.NUGGETS).addTag(NUGGETS_STEEL);
 
         this.tag(PLATES_COPPER).add(COPPER_PLATE.get());
         this.tag(PLATES_IRON).add(IRON_PLATE.get());
         this.tag(PLATES_GOLD).add(GOLD_PLATE.get());
+        this.tag(PLATES_STEEL).add(STEEL_PLATE.get());
 
         this.tag(PLATES).addTag(PLATES_COPPER);
         this.tag(PLATES).addTag(PLATES_IRON);
         this.tag(PLATES).addTag(PLATES_GOLD);
+        this.tag(PLATES).addTag(PLATES_STEEL);
 
         this.tag(GEARS_WOODEN).add(WOODEN_GEAR.get());
         this.tag(GEARS_COPPER).add(COPPER_GEAR.get());
         this.tag(GEARS_IRON).add(IRON_GEAR.get());
         this.tag(GEARS_GOLD).add(GOLD_GEAR.get());
+        this.tag(GEARS_STEEL).add(STEEL_GEAR.get());
 
         this.tag(GEARS).addTag(GEARS_WOODEN);
         this.tag(GEARS).addTag(GEARS_COPPER);
         this.tag(GEARS).addTag(GEARS_IRON);
         this.tag(GEARS).addTag(GEARS_GOLD);
+        this.tag(GEARS).addTag(GEARS_STEEL);
 
         this.tag(RODS_COPPER).add(COPPER_ROD.get());
         this.tag(RODS_IRON).add(IRON_ROD.get());
         this.tag(RODS_GOLD).add(GOLD_ROD.get());
+        this.tag(RODS_STEEL).add(STEEL_ROD.get());
 
         this.tag(Tags.Items.RODS).addTag(RODS_COPPER);
         this.tag(Tags.Items.RODS).addTag(RODS_IRON);
         this.tag(Tags.Items.RODS).addTag(RODS_GOLD);
+        this.tag(Tags.Items.RODS).addTag(RODS_STEEL);
     }
 }